requirements.txt
Django==3.2.3
django-extensions==3.1.3
djangorestframework==3.12.4
channels==3.0.3
channels_redis==3.2.0
django-createsuperuserwithpassword==2.0.0
whitenoise==5.2.0
dj-database-url==0.5.0
psycopg2-binary==2.8.6
django-constance[database]==2.8.0
django-picklefield==3.0.1